+++ servers/java/coweb-server/src/main/java/org/coweb/DefaultSessionModerator.java
@@ -129,4 +129,9 @@ public class DefaultSessionModerator extends SessionModerator {
 		return;
 	}
 
+	@Override
+	public void onSessionReady() {
+	}
+
 }
+