+++ docs/changelog.rst
@@ -17,6 +17,10 @@ Version 0.8.4
   Now, onServiceResponse is called when a bot responds to a private message
   sent by the moderator. This method is not called when other clients receive
   private bot message responses.
+* SessionModerator now provides an onReady() callback when the moderator has
+  been initialized. The motivating factor for providing this is the moderator
+  is not ready to create CollabInterface objects right after moderator
+  construction.
 
 Version 0.8.3.1
 ~~~~~~~~~~~~~~~